% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{as_odk}
\alias{as_odk}
\title{create ODK object}
\usage{
as_odk(x, questions, choices, ...)
}
\arguments{
\item{x}{a data.frame or tibble with the recorded surveys}

\item{questions}{the survey sheet in the kobo questionnaire as a data.frame}

\item{choices}{the choices sheet in the kobo questionnaire as a data.frame}

\item{...}{(ignored)}
}
\description{
create ODK object
}
